+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
 ### Fixed
+- Incorrect booking timezone used for displaying booking info in the cart and confirmation page.
 - Fatal error for undefined method call when calculating a cart booking's price.
+
+### Changed
+- Improved booking timezone fallback mechanism with better error handling and reporting.
 
 ## [0.1-alpha6] - 2018-08-01
 ### Added